+++ CHANGELOG.md
@@ -100,6 +100,7 @@
 68. Temporarily Disabled Start Screen on Unity Versions lower than 2018.
 69. Improved Exception Catching on Start Screen.
 70. Added '{' and '}' to Assembly Verifier whitelist.
+71. Fixed Issue with Melon Author Coloring.
 
 ---